feat: Added support for building native images for Drifty CLI and GUI in the new modular Drifty
