+++ .github/workflows/build.yml
@@ -2,19 +2,15 @@ name: Build and Package Drifty
 
 on:
   push:
-    branches: [ "master" ]
     paths-ignore:
       - "Website/**"
       - "*.md"
       - "*.txt"
-      - "qodana.yaml"
   pull_request:
-    branches: [ "master" ]
     paths-ignore:
       - "Website/**"      
       - "*.md"
       - "*.txt"
-      - "qodana.yaml"
   workflow_dispatch:
      inputs:
         create_release:
@@ -47,10 +43,10 @@ jobs:
       run: |
         sudo apt-get install libasound2-dev libavcodec-dev libavformat-dev libavutil-dev libfreetype6-dev
         sudo apt-get install libgl-dev libglib2.0-dev libgtk-3-dev libpango1.0-dev libx11-dev libxtst-dev zlib1g-dev
-    - name: Set up GraalVM JDK 20.0.1
+    - name: Set up GraalVM JDK 21
       uses: graalvm/setup-graalvm@v1
       with:
-        java-version: '20.0.1'
+        java-version: '21'
         distribution: 'graalvm-community'
         github-token: ${{ secrets.GITHUB_TOKEN }}
         set-java-home: true
@@ -58,7 +54,7 @@ jobs:
         components: 'native-image'
     - name: Package Drifty CLI for ${{ matrix.os }} with GraalVM
       if: ${{ matrix.mode == 'CLI' }}
-      run: mvn -Pbuild-drifty-cli-for-${{ matrix.os }} package
+      run: mvn -P build-drifty-cli-for-${{ matrix.os }} package
     - name: Set Up Maven version 3.8.8 # For GUI build issues, maven version 3.8.8 needs to be used
       if: ${{ matrix.mode == 'GUI' }}
       uses: stCarolas/setup-maven@v4.5
@@ -69,48 +65,48 @@ jobs:
       run: gcc -c config/missing_symbols.c -o config/missing_symbols-${{ matrix.os }}.o
     - name: Package Drifty GUI for ${{ matrix.os }} with GluonFX maven plugin
       if: ${{ matrix.mode == 'GUI' }}
-      run: mvn -Pbuild-drifty-gui-for-${{ matrix.os }} gluonfx:build gluonfx:package
+      run: mvn -P build-drifty-gui-for-${{ matrix.os }} gluonfx:build gluonfx:package -rf :GUI
     - name: Create Application Type specific folders
       run: |
-        mkdir ${{ matrix.mode }}
+        mkdir build/${{ matrix.mode }}
     - name: Categorise build artifacts for linux
       if: ${{ matrix.os == 'ubuntu-latest' }}
       run: |
         if ${{ matrix.mode == 'CLI' }}; then
-            mv "target/CLI/linux/Drifty CLI" "target/CLI/linux/Drifty-CLI_linux"
-            mv "target/CLI/linux/Drifty-CLI_linux" -t CLI
+            mv "CLI/target/CLI/linux/Drifty CLI" "CLI/target/CLI/linux/Drifty-CLI_linux"
+            mv "CLI/target/CLI/linux/Drifty-CLI_linux" -t build/CLI
         else
-            mv target/gluonfx/x86_64-linux/Drifty "target/gluonfx/x86_64-linux/Drifty-GUI_linux"
-            mv "target/gluonfx/x86_64-linux/Drifty-GUI_linux" -t GUI
+            mv GUI/target/gluonfx/x86_64-linux/GUI "GUI/target/gluonfx/x86_64-linux/Drifty-GUI_linux"
+            mv "GUI/target/gluonfx/x86_64-linux/Drifty-GUI_linux" -t build/GUI
         fi
     - name: Categorise build artifacts for windows
       if: ${{ matrix.os == 'windows-latest' }}
       shell: bash
       run: |
         if ${{ matrix.mode == 'CLI' }}; then
-            mv "target/CLI/windows/Drifty CLI.exe" "target/CLI/windows/Drifty-CLI.exe"
-            mv "target/CLI/windows/Drifty-CLI.exe" CLI
+            mv "CLI/target/CLI/windows/Drifty CLI.exe" "CLI/target/CLI/windows/Drifty-CLI.exe"
+            mv "CLI/target/CLI/windows/Drifty-CLI.exe" build/CLI
         else
-            mv target/gluonfx/x86_64-windows/Drifty-1.0.msi "target/gluonfx/x86_64-windows/Drifty-GUI.msi"
-            mv target/gluonfx/x86_64-windows/Drifty.exe "target/gluonfx/x86_64-windows/Drifty-GUI.exe"
-            mv "target/gluonfx/x86_64-windows/Drifty-GUI.msi" GUI
-            mv "target/gluonfx/x86_64-windows/Drifty-GUI.exe" GUI
+            mv GUI/target/gluonfx/x86_64-windows/GUI-1.0.msi "GUI/target/gluonfx/x86_64-windows/Drifty-GUI.msi"
+            mv GUI/target/gluonfx/x86_64-windows/GUI.exe "GUI/target/gluonfx/x86_64-windows/Drifty-GUI.exe"
+            mv "GUI/target/gluonfx/x86_64-windows/Drifty-GUI.msi" build/GUI
+            mv "GUI/target/gluonfx/x86_64-windows/Drifty-GUI.exe" build/GUI
         fi
     - name: Categorise build artifacts for macOS
       if: ${{ matrix.os == 'macos-latest' }}
       run: |
         if ${{ matrix.mode == 'CLI' }}; then
-            mv "target/CLI/macos/Drifty CLI" "target/CLI/macos/Drifty-CLI_macos"
-            mv "target/CLI/macos/Drifty-CLI_macos" CLI
+            mv "CLI/target/CLI/macos/Drifty CLI" "CLI/target/CLI/macos/Drifty-CLI_macos"
+            mv "CLI/target/CLI/macos/Drifty-CLI_macos" build/CLI
         else
-            mv target/gluonfx/x86_64-darwin/Drifty-1.0.0.pkg "target/gluonfx/x86_64-darwin/Drifty-GUI.pkg"
-            mv target/gluonfx/x86_64-darwin/Drifty.app "target/gluonfx/x86_64-darwin/Drifty-GUI.app"
-            mv "target/gluonfx/x86_64-darwin/Drifty-GUI.pkg" GUI
-            mv "target/gluonfx/x86_64-darwin/Drifty-GUI.app" GUI
+            mv GUI/target/gluonfx/x86_64-darwin/GUI-1.0.0.pkg "GUI/target/gluonfx/x86_64-darwin/Drifty-GUI.pkg"
+            mv GUI/target/gluonfx/x86_64-darwin/GUI.app "GUI/target/gluonfx/x86_64-darwin/Drifty-GUI.app"
+            mv "GUI/target/gluonfx/x86_64-darwin/Drifty-GUI.pkg" build/GUI
+            mv "GUI/target/gluonfx/x86_64-darwin/Drifty-GUI.app" build/GUI
         fi
     - name: Tar build files # To prevent file permission loss
       run: |
-        tar -cvf ${{ matrix.os }}-${{ matrix.mode }}.tar ${{ matrix.mode }}
+        tar -cvf ${{ matrix.os }}-${{ matrix.mode }}.tar build/${{ matrix.mode }}
     - name: Push generated artifacts for ${{ matrix.os }} for ${{ matrix.mode }}
       uses: actions/upload-artifact@v4
       with:
