+++ .github/workflows/docker-publish.yml
@@ -1,6 +1,6 @@
 name: Build and publish Docker images
 
-# This workflow uses actions that are not certified by GitHub.
+# This workflow uses actions that GitHub does not certify.
 # They are provided by a third-party and are governed by
 # separate terms of service, privacy policy, and support
 # documentation.
@@ -14,13 +14,11 @@ on:
       - "Website/**"
       - "*.md"
       - "*.txt"
-      - "qodana.yaml"
   pull_request:
     paths-ignore:
       - "Website/**"
       - "*.md"
       - "*.txt"
-      - "qodana.yaml"
   workflow_dispatch:
 
 env:
@@ -71,10 +69,10 @@ jobs:
         run: |
           sudo apt-get install libasound2-dev libavcodec-dev libavformat-dev libavutil-dev libfreetype6-dev
           sudo apt-get install libgl-dev libglib2.0-dev libgtk-3-dev libpango1.0-dev libx11-dev libxtst-dev zlib1g-dev
-      - name: Set up GraalVM JDK 20.0.1
+      - name: Set up GraalVM JDK 21
         uses: graalvm/setup-graalvm@v1
         with:
-          java-version: '20.0.1'
+          java-version: '21'
           distribution: 'graalvm-community'
           github-token: ${{ secrets.GITHUB_TOKEN }}
           set-java-home: true
@@ -82,7 +80,7 @@ jobs:
           components: 'native-image'
       - name: Package Drifty CLI with GraalVM
         if: matrix.filename == 'CLI/Dockerfile'
-        run: mvn -Pbuild-drifty-cli-for-ubuntu-latest package
+        run: mvn -P build-drifty-cli-for-ubuntu-latest package
       - name: Set Up Maven version 3.8.8 # For GUI build issues, maven version 3.8.8 needs to be used
         if: matrix.filename == 'GUI/Dockerfile'
         uses: stCarolas/setup-maven@v4.5
@@ -93,19 +91,19 @@ jobs:
         run: gcc -c config/missing_symbols.c -o config/missing_symbols-ubuntu-latest.o
       - name: Package Drifty GUI with GraalVM
         if: matrix.filename == 'GUI/Dockerfile'
-        run: mvn -Pbuild-drifty-gui-for-ubuntu-latest gluonfx:build gluonfx:package
+        run: mvn -P build-drifty-gui-for-ubuntu-latest gluonfx:build gluonfx:package -rf :GUI
       - name: Categorise build artifacts for CLI
         if: matrix.filename == 'CLI/Dockerfile'
         run: |
-          mkdir CLI
-          mv "target/CLI/linux/Drifty CLI" "target/CLI/linux/Drifty_CLI"
-          mv "target/CLI/linux/Drifty_CLI" -t CLI
+          mkdir build/CLI
+          mv "CLI/target/CLI/linux/Drifty CLI" "CLI/target/CLI/linux/Drifty_CLI"
+          mv "CLI/target/CLI/linux/Drifty_CLI" -t build/CLI
       - name: Categorise build artifacts for GUI
         if: matrix.filename == 'GUI/Dockerfile'
         run: |
-          mkdir GUI
-          mv target/gluonfx/x86_64-linux/Drifty "target/gluonfx/x86_64-linux/Drifty_GUI"
-          mv "target/gluonfx/x86_64-linux/Drifty_GUI" -t GUI
+          mkdir build/GUI
+          mv GUI/target/gluonfx/x86_64-linux/GUI "GUI/target/gluonfx/x86_64-linux/Drifty_GUI"
+          mv "GUI/target/gluonfx/x86_64-linux/Drifty_GUI" -t build/GUI
 
       - name: Set up QEMU
         uses: docker/setup-qemu-action@v3
