+++ Core/pom.xml
@@ -44,10 +44,5 @@
             <artifactId>jproc</artifactId>
             <version>2.8.2</version>
         </dependency>
-        <dependency>
-            <groupId>org.openjfx</groupId>
-            <artifactId>javafx-controls</artifactId>
-            <version>21.0.1</version>
-        </dependency>
     </dependencies>
 </project>
