+++ Core/src/main/java/properties/OS.java
@@ -31,6 +31,9 @@ public enum OS {
     }
 
     public static String getOSName() {
+        if (osName == null) {
+            setOSType();
+        }
         return osName;
     }
