+++ GUI/src/main/java/gui_support/Colors.java
@@ -1,0 +1,13 @@
+package gui_support;
+
+import javafx.scene.paint.Color;
+
+public final class Colors {
+    public static final Color GREEN = Color.rgb(0, 150, 0);
+    public static final Color RED = Color.rgb(157, 0, 0);
+    public static final Color PURPLE = Color.rgb(125, 0, 75);
+    public static final Color HOTPINK = Color.rgb(255, 0, 175);
+    public static final Color BLACK = Color.rgb(0, 0, 0);
+    public static final Color YELLOW = Color.rgb(255, 255, 0);
+    public static final Color BLUE = Color.rgb(0, 100, 255);
+}
