+++ GUI/src/main/resources/CSS/Button.css
@@ -1,0 +1,20 @@
+.button {
+    -fx-background-color: linear-gradient(rgb(54,151,225) 18%, rgb(121,218,232) 90%, rgb(126,223,255) 95%);
+    -fx-border-color: black;
+}
+
+.button:pressed {
+    -fx-background-color: linear-gradient(rgb(126,223,255) 20%, rgb(121,218,232) 20%, rgb(54,151,225) 100%);
+
+}
+
+.glassButton {
+    -fx-background-color: transparent;
+    -fx-border-color: transparent;
+}
+
+.glassButton:pressed {
+    -fx-background-color: transparent;
+    -fx-border-color: transparent;
+}
+
