+++ GUI/src/main/resources/CSS/CheckBox.css
@@ -1,0 +1,11 @@
+.check-box {
+    -fx-background-color: transparent;
+    -fx-background-image: url('CheckBoxBorder.png');
+    -fx-background-repeat: stretch;
+    -fx-background-size: 99% 99%;
+    -fx-padding: 0px;
+    -fx-font-size: 20;
+}
+.check-box:determinate .box, .check-box:determinate {
+    -fx-background-color: transparent;
+}
