+++ GUI/src/main/resources/CSS/ContextMenu.css
@@ -1,0 +1,7 @@
+.rightClick {
+    -fx-background-color: rgb(200,200,200);
+    -fx-border-color: rgba(200,200,200,1.0);
+    -fx-border-width: 1px;
+    -fx-border-radius: 5px;
+    -fx-padding: 5px;
+}
