+++ GUI/src/main/resources/CSS/Label.css
@@ -1,0 +1,16 @@
+.normalLabel {
+    -fx-font-family: Arial;
+    -fx-font-weight: bold;
+    -fx-font-size: 20;
+    -fx-text-fill: ghostwhite
+}
+.anchor {
+    -fx-font-family: Helvetica;
+    -fx-font-weight: bold;
+    -fx-font-size: 30;
+    -fx-text-fill:rgb(0,103,193)
+}
+.outline.label .text {
+    -fx-stroke: rgb(50,50,50);
+    -fx-stroke-width: .02em;
+}
