+++ GUI/src/main/resources/CSS/ListView.css
@@ -1,0 +1,35 @@
+.list-view {
+    -fx-background-color: transparent;
+    -fx-background-image: url('ListViewBorder.png');
+    -fx-background-repeat: stretch;
+    -fx-background-size: 100% 100%;
+    -fx-padding: 5px;
+}
+.list-cell {
+    -fx-background-color: transparent;
+    -fx-text-fill: rgb(40,40,40);
+}
+.list-cell:hover {
+    -fx-background-color: rgb(9,106,216);
+    -fx-text-fill: ghostwhite;
+}
+.list-view .scroll-bar:horizontal,
+.list-view .scroll-bar:vertical {
+    -fx-background-color: transparent;
+    -fx-border-color: transparent;
+}
+.list-view .scroll-bar:horizontal .thumb,
+.list-view .scroll-bar:vertical .thumb {
+    -fx-background-color:rgba(255,255,255,.15);
+    -fx-background-insets: 0, 0, 0;
+    -fx-background-radius: 0em;
+}
+.list-view .increment-button ,.list-view .decrement-button {
+    -fx-background-color:transparent;
+    -fx-border-color:transparent;
+}
+.list-view .scroll-bar:horizontal .track ,
+.list-view .scroll-bar:vertical .track{
+    -fx-background-color: transparent;
+    -fx-border-color:transparent;
+}
