+++ GUI/src/main/resources/CSS/Menu.css
@@ -1,0 +1,35 @@
+
+.menu-bar .menu .menu-item:hover {
+    -fx-background-color: #0046e5;
+    -fx-text-fill: black;
+}
+.menu:hover, .menu:showing {
+    -fx-background-color: #0046e5;
+    -fx-text-fill: black;
+}
+.menu .label:hover {
+    -fx-text-fill: ghostwhite;
+}
+.menu-item {
+    /*-fx-background-color: #F2F2F2;*/
+    -fx-text-fill: #000000;
+    -fx-font-family: verdana;
+    -fx-font-size: 11.4px;
+    -fx-background-color: ghostwhite;
+    -fx-border-color: rgba(200,200,200,1);
+    -fx-border-width: 0px;
+}
+.menu-item .label {
+    -fx-text-fill: #000000;
+    -fx-font-family: verdana;
+}
+.menu-item:focused .label {
+    -fx-text-fill: ghostwhite;
+    -fx-font-family: verdana;
+}
+.menu-item:focused {
+    -fx-background: -fx-accent;
+    -fx-background-color: #0046e5;
+    -fx-text-fill: #000000;
+    -fx-font-family: verdana;
+}
