+++ GUI/src/main/resources/CSS/Scene.css
@@ -1,0 +1,4 @@
+.root {
+    /*-fx-background-color: linear-gradient(rgb(179, 182, 182) 20%, rgb(125, 130, 130) 65%, rgb(102, 106, 106) 100%);*/
+    -fx-background-color: linear-gradient(rgba(54,151,225,1) 18%, rgba(121,218,232,1) 90%, rgba(126,223,255,1) 95%);
+}
