+++ GUI/src/main/resources/CSS/ScrollPane.css
@@ -1,0 +1,3 @@
+.scroll-pane .viewport {
+    -fx-background-color: transparent;
+}
