+++ GUI/src/main/resources/CSS/VBox.css
@@ -1,0 +1,9 @@
+.vbox {
+    -fx-background-color: rgba(0,0,0,.75);
+    -fx-border-color: rgba(255,255,255,.5);
+    -fx-border-width: 2px;
+    -fx-border-radius: 5px;
+}
+.text {
+    -fx-background-color: rgba(0,0,0,.75);-fx-border-color: rgba(255,255,255,.5);-fx-border-width: 2px;-fx-border-radius: 5px;
+}
