+++ GUI/src/main/resources/META-INF/native-image/predefined-classes-config.json
@@ -1,0 +1,8 @@
+[
+  {
+    "type":"agent-extracted",
+    "classes":[
+    ]
+  }
+]
+
