+++ GUI/src/main/resources/META-INF/native-image/resource-config.json
@@ -1,0 +1,136 @@
+{
+  "resources":{
+  "includes":[{
+    "pattern":"GUI:\\QBackgrounds/DriftyMain.png\\E"
+  }, {
+    "pattern":"GUI:\\QButtons/Save/SaveDown.png\\E"
+  }, {
+    "pattern":"GUI:\\QButtons/Save/SaveUp.png\\E"
+  }, {
+    "pattern":"GUI:\\QButtons/Start/StartDown.png\\E"
+  }, {
+    "pattern":"GUI:\\QButtons/Start/StartUp.png\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/Button.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/CheckBox.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/ContextMenu.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/Label.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/ListView.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/Menu.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/ProgressBar.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/Scene.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/ScrollPane.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/TextField.css\\E"
+  }, {
+    "pattern":"GUI:\\QCSS/VBox.css\\E"
+  }, {
+    "pattern":"GUI:\\QFonts/Monaco.ttf\\E"
+  }, {
+    "pattern":"GUI:\\QIcons/AppIcon.png\\E"
+  }, {
+    "pattern":"GUI:\\QLabels/AutoPaste.png\\E"
+  }, {
+    "pattern":"GUI:\\QLabels/Directory.png\\E"
+  }, {
+    "pattern":"GUI:\\QLabels/Filename.png\\E"
+  }, {
+    "pattern":"GUI:\\QLabels/Link.png\\E"
+  }, {
+    "pattern":"GUI:\\QSplash.png\\E"
+  }, {
+    "pattern":"\\QMETA-INF/fonts.mf\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/java.lang.System$LoggerFinder\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/java.net.spi.InetAddressResolverProvider\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/java.net.spi.URLStreamHandlerProvider\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/java.nio.charset.spi.CharsetProvider\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/java.time.zone.ZoneRulesProvider\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/java.util.prefs.PreferencesFactory\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/javax.xml.parsers.DocumentBuilderFactory\\E"
+  }, {
+    "pattern":"\\QMETA-INF/services/javax.xml.transform.TransformerFactory\\E"
+  }, {
+    "pattern":"\\Qspotdl_linux\\E"
+  }, {
+    "pattern":"\\Qyt-dlp\\E"
+  }, {
+    "pattern":"java.base:\\Qjdk/internal/icu/impl/data/icudt72b/nfkc.nrm\\E"
+  }, {
+    "pattern":"java.base:\\Qjdk/internal/icu/impl/data/icudt72b/ubidi.icu\\E"
+  }, {
+    "pattern":"java.base:\\Qjdk/internal/icu/impl/data/icudt72b/uprops.icu\\E"
+  }, {
+    "pattern":"java.base:\\Qsun/net/idn/uidna.spp\\E"
+  }, {
+    "pattern":"java.xml:\\Qcom/sun/org/apache/xml/internal/serializer/Encodings.properties\\E"
+  }, {
+    "pattern":"javafx.controls:\\Qcom/sun/javafx/scene/control/skin/modena/modena.bss\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/glass/utils/NativeLibLoader.class\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/DrawPgram_Color.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/FillPgram_Color.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/FillPgram_LinearGradient_PAD.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/FillRoundRect_Color.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/Mask_TextureSuper.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/Solid_TextureFirstPassLCD.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/Solid_TextureRGB.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/Solid_TextureSecondPassLCD.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/Texture_Color.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/prism/es2/glsl/Texture_LinearGradient_PAD.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/scenario/effect/impl/es2/glsl/Blend_SRC_IN.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/scenario/effect/impl/es2/glsl/LinearConvolveShadow_20.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/scenario/effect/impl/es2/glsl/LinearConvolveShadow_28.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qcom/sun/scenario/effect/impl/es2/glsl/LinearConvolveShadow_4.frag\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qlibglass.so\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qlibglassgtk3.so\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qlibjavafx_font.so\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qlibjavafx_font_freetype.so\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qlibjavafx_font_pango.so\\E"
+  }, {
+    "pattern":"javafx.graphics:\\Qlibprism_es2.so\\E"
+  }]},
+  "bundles":[{
+    "name":"com.sun.javafx.tk.quantum.QuantumMessagesBundle",
+    "locales":["", "und"]
+  }, {
+    "name":"com.sun.org.apache.xml.internal.serializer.XMLEntities",
+    "locales":["", "und"]
+  }, {
+    "name":"com/sun/javafx/scene/control/skin/resources/controls",
+    "locales":["", "und"]
+  }]
+}
