+++ GUI/src/main/resources/META-INF/native-image/serialization-config.json
@@ -1,0 +1,8 @@
+{
+  "types":[
+  ],
+  "lambdaCapturingTypes":[
+  ],
+  "proxies":[
+  ]
+}
